refactor: Introduce HttpCore for improved service structure
